requirements.txt
# Core Framework
playwright>=1.40.0
docling>=2.66.0
networkx>=3.0
pydantic>=2.0
fastapi
uvicorn
python-dotenv

# Web Scraping & Parsing
lxml
beautifulsoup4
requests

# Vector Database (existing, now explicit)
chromadb>=0.4.0

# === GRAPH RAG ENHANCEMENTS ===

# Phase 1: Sparse Retrieval & Reranking
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# German NLP
spacy>=3.5.0

# Phase 2: Graph Algorithms
python-louvain>=0.16

# Scientific Computing
scipy>=1.10.0
numpy>=1.24.0

# Phase 6: Evaluation
scikit-learn>=1.3.0
rouge-score>=0.1.2

# Optional: Better performance
# psutil>=5.9.0  # Memory monitoring
